feat: support playing again after ended
If .play() or a seek is initiated after ended,
treat it as a new session and emit loaded.

Browsers handle this slightly differently, pressing
play when at end of stream is at least partially
treated as a seek, but the filter will now treat it
as a new playback session and trigger LOADED when video
is ready to start playing again.